Restructured FindRemoteFiles -> RetrieveIrodsCollection.
